pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[project]
name = "ironclad-mcp"
version = "0.1.0"
description = "MCP server for querying Ironclad contract data with Vault integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.1.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.27.0",
    "uvicorn>=0.24.0",
    "google-cloud-secret-manager>=2.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
]

[project.scripts]
ironclad-mcp = "ironclad_mcp.server:main"
ironclad-mcp-http = "ironclad_mcp.http_server:main"

















